Pdf.Core.Object.Builder

Description

Render Object to bytestring

Documentation

buildIndirectObject :: Ref -> Object -> Builder

Build indirect object except streams

buildIndirectStream :: Ref -> Dict -> ByteString -> Builder

Build indirect stream

buildObject :: Object -> Builder

Render inline object (without "obj/endobj"). It is error to supply Stream, because it could not be inlined, but should always be an indirect object

buildNumber :: Scientific -> Builder

Build a number

buildBool :: Bool -> Builder

Build a bool

buildName :: Name -> Builder

Build a name

buildDict :: Dict -> Builder

Build a dictionary

buildArray :: Array -> Builder

Build an array

buildString :: ByteString -> Builder

Build a string

It may produce literal or hex string based on the context.

buildRef :: Ref -> Builder

Build a reference

buildStream :: Dict -> ByteString -> Builder

Build a stream

The function doesn't try to encode or encrypt the content