Pdf.Core.Parsers.Object

Description

This module contains parsers for pdf objects

Parse any object

parseObject :: Parser Object

Parse any Object except Stream because for Stream we need offset of data in file

>>> parseOnly parseObject "/Name"
Right (OName (Name "Name"))

Parse object of specific type

parseDict :: Parser Dict

Parse a dictionary

parseArray :: Parser Array

Parse an array

parseName :: Parser Name

Parse a name

parseString :: Parser ByteString

Parse literal string

parseHexString :: Parser ByteString

Parse hex string

parseRef :: Parser Ref

Parse a reference

parseNumber :: Parser Scientific

Parse number

parseBool :: Parser Bool

Parse bool value

Other

parseTillStreamData :: Parser ()

Consumes input till stream's data

Use parseDict then parseTillStreamData to determine whether the object is dictionary or stream. If parseTillStreamData fails, then it is a dictionary. Otherwise it is stream, and current position in input data will point to stream's data start

>>> parse (parseDict >>= \dict -> parseTillStreamData >> return dict) "<</Key 123>>\nstream\n1234\nendstream"
Done "1234\nendstream" Dict [(Name "Key",ONumber (NumInt 123))]

parseIndirectObject :: Parser (Ref, Object)

Parse object. Input position should point to offset defined in XRef

>>> parseOnly parseIndirectObject "1 2 obj\n12"
Right (Ref 1 2,ONumber (NumInt 12))

isRegularChar :: Char -> Bool

Whether the character can appear in Name