Pdf.Core.Writer

Description

Write PDF files

It could be used to generate new PDF file or to incrementally update the existent one

To generate new file, first call writeHeader, then a number of writeObject and finally writeXRefTable

To incrementally update PDF file just omit the writeHeader and append the result to the existent file

Documentation

data Writer

makeWriter :: OutputStream ByteString -> IO Writer

writeHeader :: Writer -> IO ()

Write PDF header. Used for generating new PDF files. Should be the first call. Not used fo incremental updates

writeObject :: Writer -> Ref -> Object -> IO ()

Write object

writeStream :: Writer -> Ref -> Dict -> ByteString -> IO ()

Write stream

deleteObject :: Writer -> Ref -> Int64 -> IO ()

Delete object

writeXRefTable

Arguments

:: Writer

-> Int64
size of the original PDF file. Should be 0 for new file

-> Dict
trailer

-> IO ()

Write xref table. Should be the last call. Used for generating and incremental updates.