Pdf.Core.XRef

Description

Cross reference

Documentation

data XRef

Cross reference

Constructors

XRefTable Int64
Offset

XRefStream Int64 Stream
Offset and stream

data Entry

Entry in cross reference stream

Constructors

EntryFree Int Int
Object number and generation

EntryUsed Int64 Int
Object offset (in bytes from the beginning of file) and generation

EntryCompressed Int Int
Object number of object stream and index within the object stream

readXRef :: Buffer -> Int64 -> IO XRef

Read XRef at specified offset

lastXRef :: Buffer -> IO XRef

Find the last cross reference

prevXRef :: Buffer -> XRef -> IO (Maybe XRef)

Find prev cross reference

trailer :: Buffer -> XRef -> IO Dict

Read trailer for the xref

lookupTableEntry

Arguments

:: Buffer

-> XRef
should be xref table

-> Ref
indirect object to look for

-> IO (Maybe Entry)

Read xref entry for the indirect object from xref table

lookupStreamEntry

Arguments

:: Dict
xref stream dictionary

-> InputStream ByteString
decoded xref stream content

-> Ref
indirect object

-> IO (Maybe Entry)

Read xref entry for the indirect object from xref stream

See pdf1.7 spec: 7.5.8 Cross-Reference Streams. May throw UnknownXRefStreamEntryType

isTable :: InputStream ByteString -> IO Bool

Check whether the stream starts with "xref" keyword. The keyword itself and newline after it are consumed

data UnknownXRefStreamEntryType

Unknown entry type should be interpreted as reference to null object

Constructors

UnknownXRefStreamEntryType Int